restructure ObstacleAvoidanceDialogs folder
