+++ src/web/components/ObstacleAvoidanceDialogs/Common/ObstacleAvoidanceBaseDialog.tsx
