+++ src/web/components/ObstacleAvoidanceDialogs/MissionRerouteDialog/MissionRerouteDialog.tsx
@@ -6,8 +6,8 @@ import {
     PendingReroute,
     ProposalStatus,
 } from "../../../data/obstacle_avoidance_data/pending-route-data";
-import ObstacleAvoidanceBaseDialog from "../ObstacleAvoidanceBaseDialog";
-import RerouteSummary from "../RerouteSummary";
+import ObstacleAvoidanceBaseDialog from "../Common/ObstacleAvoidanceBaseDialog";
+import RerouteSummary from "../Common/RerouteSummary";
 
 export default function MissionRerouteDialog({ pending }: { pending: PendingReroute }) {
     const jaiaDispatch = useContext(JaiaDispatchContext);
