+++ src/web/components/ObstacleAvoidanceDialogs/docs/ADD_MOVE_WAYPOINT_CONSISTENCY.md
