+++ src/web/components/ObstacleAvoidanceDialogs/docs/PENDING_DIALOG_REFACTOR_PLAN.md
