+++ src/web/components/ObstacleAvoidanceDialogs/docs/REFACTOR_PLAN.md
